requirements.txt
Flask
pandas
requests
python-dotenv
openpyxl
discord-webhook